--- src/features/Login/types.ts
@@ -1,8 +1,0 @@
-export interface LoginRequest {
-  email: string;
-  password: string;
-}
-
-export interface LoginSuccessResponse {
-  accessToken: string;
-}
